requirements.txt
requests~=2.33.1
allure-python-commons~=2.15.3
pytest~=9.0.2
Flask~=3.1.3
PyMySQL~=1.1.2
PyYAML~=6.0.3